\name{getlevels}
\alias{getlevels}
\alias{getlevels-methods}
\title{Levels frequency}
\description{Gives the calls status at every position}
\usage{getlevels(.Object)}
\arguments{
  \item{.Object}{an object of the CGHresults class}
}
\value{
  \item{levelslist}{list of dataframes containing calls for each position}
}

\references{Joint segmentation, calling and normalization of multiple CGH profiles, Biostatistics (2011)}
\author{F. Picard, E. Lebarbier, M. Hoebeke, G. Rigaill, B. Thiam, S. Robin}
\keyword{methods}

