\name{nblevels<-}
\alias{nblevels<-}
\alias{nblevels<--methods}
\title{Replace slot nblevels}
\description{changes slot \code{nblevels} of a CGHoptions object}
\arguments{
  \item{.Object}{an object of the CGHoptions class}
}
\details{ \code{CGHo["nblevels"]} determines the total number of levels
  when calling is performed.}
\references{Joint segmentation, calling and normalization of multiple CGH profiles, Biostatistics (2011)}
\author{F. Picard, E. Lebarbier, M. Hoebeke, G. Rigaill, B. Thiam, S. Robin}
\keyword{ts}
\keyword{methods}
